requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
pandas==2.1.4
numpy>=1.24.0
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
loguru==0.7.2
tushare==1.2.89
akshare==1.12.0
ta-lib==0.4.28
redis==5.0.1
celery==5.3.4
pytest==7.4.3
httpx==0.25.2 